+++ cmd/customPackage.go
@@ -9,70 +9,73 @@ import (
 )
 
 func init() {
-	// LVL CUSTOMPACKAGES
+	// LVL PACKAGE
 	RootCmd.AddCommand(cmdCustomPackages)
 
-	// LVL CUSTOMPACKAGES BASE
+	// LVL PACKAGE BASE
 	cmdCustomPackages.AddCommand(cmdCustomPackagesBase)
 
-	// LVL CUSTOMPACKAGES BASE GET
+	// LVL PACKAGE BASE GET
 	cmdCustomPackagesBase.AddCommand(cmdCustomPackagesBaseGet)
 
-	// LVL CUSTOMPACKAGES CREATE
+	// LVL PACKAGE CREATE
 	cmdCustomPackages.AddCommand(customPackagesCreateCmd)
-	customPackagesCreateCmd.Flags().StringVarP(&customPackagesCreateBase, "base", "b", "", "")
-	customPackagesCreateCmd.Flags().StringVarP(&customPackagesCreateName, "name", "n", "", "")
-	customPackagesCreateCmd.Flags().StringVar(&customPackagesCreateOrganisation, "organisation", "", "")
+	customPackagesCreateCmd.Flags().StringVarP(&customPackagesCreateBase, "base", "b", "", "The base package to use. Use 'lvl package base get' to show base packages.")
+	customPackagesCreateCmd.Flags().StringVarP(&customPackagesCreateName, "name", "n", "", "The name that the created package will have.")
+	customPackagesCreateCmd.Flags().StringVar(&customPackagesCreateOrganisation, "organisation", "", "Organisation that the package will be created on. Defaults to your current organisation.")
 	_ = customPackagesCreateCmd.MarkFlagRequired("base")
 	_ = customPackagesCreateCmd.MarkFlagRequired("name")
 
-	// LVL CUSTOMPACKAGES
+	// LVL PACKAGE
 	cmdCustomPackages.AddCommand(customPackagesGetCmd)
 	addCommonGetFlags(customPackagesGetCmd)
 
-	// LVL CUSTOMPACKAGES DESCRIBE
+	// LVL PACKAGE DESCRIBE
 	cmdCustomPackages.AddCommand(customPackagesDescribeCmd)
 
-	// LVL CUSTOMPACKAGES INSTANTIATE
+	// LVL PACKAGE INSTANTIATE
 	cmdCustomPackages.AddCommand(customPackageInstantiateCmd)
 	customPackageInstantiateCmd.Flags().StringVar(&customPackageInstantiateOrganisation, "organisation", "", "")
 	customPackageInstantiateCmd.Flags().StringArrayVarP(&customPackageInstantiateParams, "param", "p", nil, "")
+	customPackageInstantiateCmd.Flags().StringVarP(&customPackageInstantiateName, "name", "n", "", "Name of the created app")
+	customPackageInstantiateCmd.MarkFlagRequired(customPackageInstantiateName)
+	addWaitFlag(customPackageInstantiateCmd)
 
-	// LVL CUSTOMPACKAGES DELETE
+	// LVL PACKAGE DELETE
 	cmdCustomPackages.AddCommand(customPackagesDeleteCmd)
 	addDeleteConfirmFlag(customPackagesDeleteCmd)
 
-	// LVL CUSTOMPACKAGES TEMPLATE
+	// LVL PACKAGE TEMPLATE
 	cmdCustomPackages.AddCommand(customPackageTemplateCmd)
 
-	// LVL CUSTOMPACKAGES TEMPLATE ADD
+	// LVL PACKAGE TEMPLATE ADD
 	customPackageTemplateCmd.AddCommand(customPackageTemplateAddCmd)
 	customPackageTemplateAddCmd.Flags().StringVarP(&customPackageTemplateAddTemplate, "template", "t", "", "")
 	customPackageTemplateAddCmd.Flags().StringVarP(&customPackageTemplateAddGroup, "group", "g", "", "")
 	_ = customPackagesCreateCmd.MarkFlagRequired("template")
 	_ = customPackagesCreateCmd.MarkFlagRequired("group")
 
-	// LVL CUSTOMPACKAGES TEMPLATE REMOVE
+	// LVL PACKAGE TEMPLATE REMOVE
 	customPackageTemplateCmd.AddCommand(customPackageTemplateRemoveCmd)
 
-	// LVL CUSTOMPACKAGES TEMPLATE TYPE
+	// LVL PACKAGE TEMPLATE TYPE
 	customPackageTemplateCmd.AddCommand(customPackageTemplateTypeCmd)
 
-	// LVL CUSTOMPACKAGES TEMPLATE TYPE GET
+	// LVL PACKAGE TEMPLATE TYPE GET
 	customPackageTemplateTypeCmd.AddCommand(customPackageTemplateTypeGetCmd)
 
-	// LVL CUSTOMPACKAGES TEMPLATE TYPE DESCRIBE
+	// LVL PACKAGE TEMPLATE TYPE DESCRIBE
 	customPackageTemplateTypeCmd.AddCommand(customPackageTemplateTypeDescribeCmd)
 }
 
 // Resolve the ID of an app based on user-provided name or ID.
-func resolveCustomerPackage(arg string) (l27.IntID, error) {
+func resolveCustomPackage(arg string) (l27.IntID, error) {
 	id, err := l27.ParseID(arg)
 	if err == nil {
 		return id, nil
 	}
 
-	options, err := Level27Client.CustomerPackageLookup(arg)
+	options, err := Level27Client.CustomPackageLookup(arg)
 	if err != nil {
 		return 0, err
 	}
@@ -81,7 +84,7 @@ func resolveCustomerPackage(arg string) (l27.IntID, error) {
 		options,
 		arg,
 		"custom package",
-		func(pack l27.CustomerPackageShort) string { return fmt.Sprintf("%s (%d)", pack.Name, pack.ID) })
+		func(pack l27.CustomPackageShort) string { return fmt.Sprintf("%s (%d)", pack.Name, pack.ID) })
 
 	if err != nil {
 		return 0, err
@@ -91,23 +94,26 @@ func resolveCustomerPackage(arg string) (l27.IntID, error) {
 }
 
 var cmdCustomPackages = &cobra.Command{
-	Use:     "custompackages",
-	Aliases: []string{"custpkgs"},
-	Hidden:  true,
+	Use:   "package",
+	Short: "Manage Agency Hosting packages",
+	Long: `Manage Agency Hosting packages.
+Packages are, fundamentally, a series of templates that get executed to create the final set of products.`,
+	Aliases: []string{"packages", "pkg", "pkgs", "custompackage", "custompackages"},
 }
 
 var cmdCustomPackagesBase = &cobra.Command{
 	Use:   "base",
-	Short: "Commands for showing base custom packages",
+	Short: "Commands for showing base packages",
 }
 
 var cmdCustomPackagesBaseGet = &cobra.Command{
-	Use:   "get",
-	Short: "List base custom packages",
+	Use:     "get",
+	Short:   "List base packages that Agency packages can be created from",
+	Example: "  lvl package base get",
 
 	Args: cobra.NoArgs,
 	RunE: func(cmd *cobra.Command, args []string) error {
-		packages, err := Level27Client.CustomPackagesGetList()
+		packages, err := Level27Client.CustPackagesGetList()
 		if err != nil {
 			return err
 		}
@@ -125,7 +131,9 @@ var customPackagesCreateBase string
 var customPackagesCreateName string
 var customPackagesCreateOrganisation string
 var customPackagesCreateCmd = &cobra.Command{
-	Use: "create -b <base package> -n <name>",
+	Use:     "create -b <base package> -n <name>",
+	Short:   "Create a new package",
+	Example: `  lvl package create -b php_mysql_generic_walk -n walk-php`,
 
 	Args: cobra.NoArgs,
 	RunE: func(cmd *cobra.Command, args []string) error {
@@ -134,13 +142,13 @@ var customPackagesCreateCmd = &cobra.Command{
 			return err
 		}
 
-		create := l27.CustomerPackageCreate{
+		create := l27.CustomPackageCreate{
 			Name:              customPackagesCreateName,
 			CustomPackageName: customPackagesCreateBase,
 			Organisation:      orgID,
 		}
 
-		pack, err := Level27Client.CustomerPackageCreate(&create)
+		pack, err := Level27Client.CustomPackageCreate(&create)
 		if err != nil {
 			return err
 		}
@@ -152,27 +160,30 @@ var customPackagesCreateCmd = &cobra.Command{
 }
 
 var customPackagesDeleteCmd = &cobra.Command{
-	Use: "delete <custom package>",
+	Use:     "delete <package name or ID>",
+	Short:   "Delete a package",
+	Long:    "Delete a package. Deleting a package does not affect created apps.",
+	Example: "  lvl package delete walk-php",
 
 	Args: cobra.ExactArgs(1),
 	RunE: func(cmd *cobra.Command, args []string) error {
-		packageID, err := resolveCustomerPackage(args[0])
+		packageID, err := resolveCustomPackage(args[0])
 		if err != nil {
 			return err
 		}
 
 		if !optDeleteConfirmed {
-			pack, err := Level27Client.CustomerPackageGetSingle(packageID)
+			pack, err := Level27Client.CustomPackageGetSingle(packageID)
 			if err != nil {
 				return err
 			}
 
-			if !confirmPrompt(fmt.Sprintf("Delete custom package %s (%d)?", pack.Name, pack.ID)) {
+			if !confirmPrompt(fmt.Sprintf("Delete package %s (%d)?", pack.Name, pack.ID)) {
 				return nil
 			}
 		}
 
-		err = Level27Client.CustomerPackageDelete(packageID)
+		err = Level27Client.CustomPackageDelete(packageID)
 		if err != nil {
 			return err
 		}
@@ -184,10 +195,35 @@ var customPackagesDeleteCmd = &cobra.Command{
 }
 
 var customPackagesGetCmd = &cobra.Command{
-	Use: "get",
+	Use:   "get [package name or ID [package...]]",
+	Short: "List packages or get info",
+	Example: `List packages:
+  lvl package get
+List packages with search:
+  lvl package get -f walk
+Get info about a specific package:
+  lvl package get walk-php
+`,
 
 	RunE: func(cmd *cobra.Command, args []string) error {
-		packages, err := Level27Client.CustomerPackageGetList(optGetParameters)
+		packages, err := resolveGets[l27.CustomPackageShort](
+			args,
+			func(s string) ([]l27.CustomPackageShort, error) {
+				return Level27Client.CustomPackageLookup(s)
+			},
+			func(i int32) (l27.CustomPackageShort, error) {
+				value, err := Level27Client.CustomPackageGetSingle(i)
+				if err != nil {
+					return l27.CustomPackageShort{}, err
+				}
+
+				return value.ToShort(), nil
+			},
+			func(cgp l27.CommonGetParams) ([]l27.CustomPackageShort, error) {
+				return Level27Client.CustomPackageGetList(cgp)
+			},
+		)
+
 		if err != nil {
 			return err
 		}
@@ -202,17 +238,19 @@ var customPackagesGetCmd = &cobra.Command{
 }
 
 var customPackagesDescribeCmd = &cobra.Command{
-	Use: "describe <custom package>",
+	Use:     "describe <package name or ID>",
+	Short:   "Show detailed information about a package",
+	Example: "  lvl package describe walk-php",
 
 	Args: cobra.ExactArgs(1),
 	RunE: func(cmd *cobra.Command, args []string) error {
 		packageName := args[0]
-		id, err := resolveCustomerPackage(packageName)
+		id, err := resolveCustomPackage(packageName)
 		if err != nil {
 			return err
 		}
 
-		pack, err := Level27Client.CustomerPackageGetSingle(id)
+		pack, err := Level27Client.CustomPackageGetSingle(id)
 		if err != nil {
 			return err
 		}
@@ -225,13 +263,18 @@ var customPackagesDescribeCmd = &cobra.Command{
 
 var customPackageInstantiateOrganisation string
 var customPackageInstantiateParams []string
+var customPackageInstantiateName string
 var customPackageInstantiateCmd = &cobra.Command{
-	Use: "instantiate <custom package>",
+	Use:   "instantiate <package name or ID> -n <name>",
+	Short: "Create an app from a package",
+	Long: `Create an app from a package.
+This only starts the creation via a task. If you wait for the task to complete with --wait,
+the resulting entities will be listed.`,
 
 	Args: cobra.ExactArgs(1),
 	RunE: func(cmd *cobra.Command, args []string) error {
 		packageName := args[0]
-		id, err := resolveCustomerPackage(packageName)
+		id, err := resolveCustomPackage(packageName)
 		if err != nil {
 			return err
 		}
@@ -242,6 +285,7 @@ var customPackageInstantiateCmd = &cobra.Command{
 		}
 
 		params := map[string]l27.ParameterValue{}
+		params["name"] = customPackageInstantiateName
 
 		for _, param := range customPackageInstantiateParams {
 			split := strings.SplitN(param, "=", 2)
@@ -255,14 +299,33 @@ var customPackageInstantiateCmd = &cobra.Command{
 			}
 		}
 
-		request := l27.CustomerPackageRootTaskRequest{
+		request := l27.CustomPackageRootTaskRequest{
 			Organisation: orgID,
 			Params:       params,
 		}
 
-		err = Level27Client.CustomerPackageRootTask(id, &request)
+		task, err := Level27Client.CustomPackageRootTask(id, &request)
 		if err != nil {
 			return err
+		}
+
+		outputFormatTemplate(task, "templates/entities/customPackages/instantiate.tmpl")
+
+		if optWait {
+			task, err = waitForStatus[l27.RootTask](
+				func() (l27.RootTask, error) {
+					return Level27Client.RootTaskGetSingle(task.Id)
+				},
+				func(rt l27.RootTask) string { return rt.Status },
+				"done",
+				[]string{"to_do", "busy"},
+			)
+
+			if err != nil {
+				return fmt.Errorf("waiting on task failed: %s", err.Error())
+			}
+
+			outputFormatTemplate(task, "templates/entities/customPackages/instantiate_full.tmpl")
 		}
 
 		return nil
@@ -270,15 +333,21 @@ var customPackageInstantiateCmd = &cobra.Command{
 }
 
 var customPackageTemplateCmd = &cobra.Command{
-	Use: "template",
+	Use:   "template",
+	Short: "Manage templates on custom packages",
+	// TODO: template add can't even add parameters so it's pretty useless right now
+	// Hiding as a result.
+	Hidden: true,
 }
 
 var customPackageTemplateTypeCmd = &cobra.Command{
-	Use: "type",
+	Use:   "type",
+	Short: "Show information for available templates",
 }
 
 var customPackageTemplateTypeGetCmd = &cobra.Command{
-	Use: "get",
+	Use:   "get",
+	Short: "List templates available for adding to an Agency package",
 
 	Args: cobra.NoArgs,
 	RunE: func(cmd *cobra.Command, args []string) error {
@@ -294,7 +363,8 @@ var customPackageTemplateTypeGetCmd = &cobra.Command{
 }
 
 var customPackageTemplateTypeDescribeCmd = &cobra.Command{
-	Use: "describe",
+	Use:   "describe <template name>",
+	Short: "Show detailed information for a package template",
 
 	Args: cobra.ExactArgs(1),
 	RunE: func(cmd *cobra.Command, args []string) error {
@@ -312,22 +382,23 @@ var customPackageTemplateTypeDescribeCmd = &cobra.Command{
 var customPackageTemplateAddTemplate string
 var customPackageTemplateAddGroup string
 var customPackageTemplateAddCmd = &cobra.Command{
-	Use: "add",
+	Use:   "add <package> -t <template>",
+	Short: "Add a template to a package",
 
 	Args: cobra.ExactArgs(1),
 	RunE: func(cmd *cobra.Command, args []string) error {
-		customPackageID, err := resolveCustomerPackage(args[0])
+		customPackageID, err := resolveCustomPackage(args[0])
 		if err != nil {
 			return err
 		}
 
-		create := l27.CustomerPackageTemplateCreate{
+		create := l27.CustomPackageTemplateCreate{
 			Template:      customPackageTemplateAddTemplate,
 			LimitGroup:    customPackageTemplateAddGroup,
 			CustomPackage: customPackageID,
 		}
 
-		template, err := Level27Client.CustomerPackageTemplateCreate(customPackageID, &create)
+		template, err := Level27Client.CustomPackageTemplateCreate(customPackageID, &create)
 		if err != nil {
 			return err
 		}
@@ -343,7 +414,7 @@ var customPackageTemplateRemoveCmd = &cobra.Command{
 
 	Args: cobra.ExactArgs(2),
 	RunE: func(cmd *cobra.Command, args []string) error {
-		customPackageID, err := resolveCustomerPackage(args[0])
+		customPackageID, err := resolveCustomPackage(args[0])
 		if err != nil {
 			return err
 		}
@@ -353,7 +424,7 @@ var customPackageTemplateRemoveCmd = &cobra.Command{
 			return err
 		}
 
-		err = Level27Client.CustomerPackageTemplateRemove(customPackageID, templateID)
+		err = Level27Client.CustomPackageTemplateRemove(customPackageID, templateID)
 		if err != nil {
 			return err
 		}